allow player to press 1 to change color only when near the box
bug* if they exit out of the collider and decided to collide back into it, the change color timeline will reset it color but not it timeline itself. ex: if you got the box to almost blue and exit the collider, when u re-enter the collider the box will turn back to pink, but if u press 1 to change it to blue, it will continue where it left off from it timeline and  since it almost blue, it just finish it off which u will see it gonna change from pink straight to blue.